Move ElementFactory to gaphor.core.modeling
