+++ gaphor/UML/__init__.py
@@ -1,7 +1,6 @@
 # Here, order matters
 from gaphor.UML.uml2 import *  # noqa: isort:skip
 from gaphor.UML.presentation import Presentation  # noqa: isort:skip
-from gaphor.UML.elementfactory import ElementFactory  # noqa: isort:skip
 from gaphor.UML import modelfactory as model  # noqa: isort:skip
 
 from gaphor.UML.umlfmt import format
