+++ gaphor/UML/tests/test_diagram.py
@@ -1,9 +1,9 @@
 import gaphas
 import pytest
 
+from gaphor.core.modeling import ElementFactory
 from gaphor.services.eventmanager import EventManager
 from gaphor.UML import Diagram
-from gaphor.UML.elementfactory import ElementFactory
 from gaphor.UML.presentation import Presentation
