+++ gaphor/UML/tests/test_modelfactory.py
@@ -1,13 +1,14 @@
 import pytest
 
 from gaphor import UML
+from gaphor.core.modeling import ElementFactory
 from gaphor.services.eventmanager import EventManager
 
 
 @pytest.fixture
 def factory():
     event_manager = EventManager()
-    return UML.ElementFactory(event_manager)
+    return ElementFactory(event_manager)
 
 
 def test_stereotype_name(factory):
