+++ gaphor/UML/tests/test_umlfmt.py
@@ -5,9 +5,9 @@ Formatting of UML model elements into text tests.
 import pytest
 
 import gaphor.UML.uml2 as UML
+from gaphor.core.modeling import ElementFactory
 from gaphor.services.eventmanager import EventManager
 from gaphor.UML import model
-from gaphor.UML.elementfactory import ElementFactory
 from gaphor.UML.umlfmt import format
 from gaphor.UML.umllex import parse
