+++ gaphor/core/modeling/__init__.py
@@ -1,0 +1,1 @@
+from gaphor.core.modeling.elementfactory import ElementFactory
