+++ gaphor/core/modeling/elementdispatcher.py
