+++ gaphor/core/modeling/elementfactory.py
@@ -17,8 +17,8 @@ from typing import (
 )
 
 from gaphor.abc import Service
+from gaphor.core.modeling.elementdispatcher import ElementDispatcher, EventWatcher
 from gaphor.UML.element import Element, UnlinkEvent
-from gaphor.UML.elementdispatcher import ElementDispatcher, EventWatcher
 from gaphor.UML.event import ElementCreated, ElementDeleted, ModelFlushed, ModelReady
 from gaphor.UML.uml2 import Diagram
