+++ gaphor/core/modeling/tests/test_elementfactory.py
@@ -4,9 +4,9 @@ import pytest
 
 from gaphor.application import Session
 from gaphor.core import event_handler
+from gaphor.core.modeling import ElementFactory
 from gaphor.services.eventmanager import EventManager
 from gaphor.UML import Parameter
-from gaphor.UML.elementfactory import ElementFactory
 from gaphor.UML.event import (
     ElementCreated,
     ElementDeleted,
