+++ gaphor/diagram/general/tests/test_propertypages.py
@@ -1,9 +1,9 @@
 import pytest
 
 from gaphor import UML
+from gaphor.core.modeling import ElementFactory
 from gaphor.diagram.general.generalpropertypages import CommentItemPropertyPage
 from gaphor.services.eventmanager import EventManager
-from gaphor.UML.elementfactory import ElementFactory
 
 
 @pytest.fixture
