+++ gaphor/diagram/tests/fixtures.py
@@ -5,11 +5,11 @@ from gaphas.aspect import ConnectionSink
 from gaphas.aspect import Connector as ConnectorAspect
 
 from gaphor import UML
+from gaphor.core.modeling import ElementFactory
 from gaphor.diagram.connectors import Connector
 from gaphor.misc.xmlwriter import XMLWriter
 from gaphor.services.eventmanager import EventManager
 from gaphor.storage import storage
-from gaphor.UML.elementfactory import ElementFactory
 
 
 @pytest.fixture
