+++ gaphor/storage/tests/test_storage_upgrades.py
@@ -1,10 +1,10 @@
 import pytest
 
+from gaphor.core.modeling import ElementFactory
 from gaphor.services.eventmanager import EventManager
 from gaphor.storage import diagramitems
 from gaphor.storage.parser import canvas, canvasitem, element
 from gaphor.storage.storage import load_elements
-from gaphor.UML.elementfactory import ElementFactory
 
 
 @pytest.fixture
