+++ gaphor/ui/tests/test_namespace.py
@@ -1,8 +1,8 @@
 import pytest
 
-import gaphor.services.componentregistry
 import gaphor.services.eventmanager
 import gaphor.UML as UML
+from gaphor.core.modeling import ElementFactory
 from gaphor.ui.namespace import Namespace
 
 
@@ -13,12 +13,7 @@ def event_manager():
 
 @pytest.fixture
 def element_factory(event_manager):
-    return UML.elementfactory.ElementFactory(event_manager)
-
-
-@pytest.fixture
-def component_registry():
-    return gaphor.services.componentregistry.ComponentRegistry()
+    return ElementFactory(event_manager)
 
 
 @pytest.fixture
