+++ pyproject.toml
@@ -70,7 +70,7 @@ gaphorconvert = 'gaphor.tools.gaphorconvert:main'
 "event_manager" = "gaphor.services.eventmanager:EventManager"
 "properties" = "gaphor.services.properties:Properties"
 "undo_manager" = "gaphor.services.undomanager:UndoManager"
-"element_factory" = "gaphor.UML.elementfactory:ElementFactory"
+"element_factory" = "gaphor.core.modeling:ElementFactory"
 "file_manager" = "gaphor.ui.filemanager:FileManager"
 "recent_files" = "gaphor.ui.recentfiles:RecentFiles"
 "main_window" = "gaphor.ui.mainwindow:MainWindow"
